// frontend/src/api/generated/models/systemDimensionalLimitOpeningTypeEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Dekopen API
 * Authenticated tenant and engine API boundary.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `FIXED` - FIXED
 * * `TURN_LEFT` - TURN_LEFT
 * * `TURN_RIGHT` - TURN_RIGHT
 * * `TILT_TURN_LEFT` - TILT_TURN_LEFT
 * * `TILT_TURN_RIGHT` - TILT_TURN_RIGHT
 * * `SLIDING_2L` - SLIDING_2L
 * * `SLIDING_3L` - SLIDING_3L
 * * `SLIDING_4L` - SLIDING_4L
 * * `SLIDING` - SLIDING
 * * `AWNING` - AWNING
 * * `DOOR_ENTRY` - DOOR_ENTRY
 * * `DOOR_DOUBLE` - DOOR_DOUBLE
 */
export type SystemDimensionalLimitOpeningTypeEnum =
  (typeof SystemDimensionalLimitOpeningTypeEnum)[keyof typeof SystemDimensionalLimitOpeningTypeEnum];

export const SystemDimensionalLimitOpeningTypeEnum = {
  FIXED: "FIXED",
  TURN_LEFT: "TURN_LEFT",
  TURN_RIGHT: "TURN_RIGHT",
  TILT_TURN_LEFT: "TILT_TURN_LEFT",
  TILT_TURN_RIGHT: "TILT_TURN_RIGHT",
  SLIDING_2L: "SLIDING_2L",
  SLIDING_3L: "SLIDING_3L",
  SLIDING_4L: "SLIDING_4L",
  SLIDING: "SLIDING",
  AWNING: "AWNING",
  DOOR_ENTRY: "DOOR_ENTRY",
  DOOR_DOUBLE: "DOOR_DOUBLE",
} as const;